src/loadgenerator/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2018.11.29       # via geventhttpclient-wheels, requests
chardet==3.0.4            # via requests
click==7.0                # via flask
faker==3.0.1              # via -r requirements.in
flask==1.0.2              # via locustio
gevent==1.4.0             # via geventhttpclient-wheels, locustio
geventhttpclient-wheels==1.3.1.dev2  # via locustio
greenlet==0.4.15          # via gevent
idna==2.8                 # via requests
ipaddress==1.0.23         # via faker
itsdangerous==1.1.0       # via flask
jinja2==2.10              # via flask
locustio==0.13.0          # via -r requirements.in
markupsafe==1.1.0         # via jinja2
msgpack-python==0.5.6     # via locustio
python-dateutil==2.8.1    # via faker
pyzmq==17.0.0             # via locustio
requests==2.21.0          # via locustio
six==1.12.0               # via faker, geventhttpclient-wheels, locustio, python-dateutil
text-unidecode==1.3       # via faker
urllib3==1.24.2           # via requests
werkzeug==0.14.1          # via flask
